+++ src/components/icons/Microsoft.tsx
@@ -1,0 +1,21 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import React from 'react'
+
+import { Icon, IconProps } from '@chakra-ui/react'
+
+export const MicrosoftIcon = (props: IconProps) => (
+  <Icon {...props}>
+    <svg
+      width='17'
+      height='16'
+      viewBox='0 0 17 16'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path d='M7.66699 0H0.166992V7.5H7.66699V0Z' fill='#F35325' />
+      <path d='M16.167 0H8.66699V7.5H16.167V0Z' fill='#81BC06' />
+      <path d='M7.66699 8.5H0.166992V16H7.66699V8.5Z' fill='#05A6F0' />
+      <path d='M16.167 8.5H8.66699V16H16.167V8.5Z' fill='#FFBA08' />
+    </svg>
+  </Icon>
+)